Split schema tests from dag tests
The main schema tests were in tests/unit/dag . Move to tests/unit/schema
to be consistent with the merlin.schema package layout.
